fix getFee bug.  Updated Unit Tests

+++ fiosdk/src/main/java/fiofoundation/io/fiosdk/models/fionetworkprovider/response/GetFeeResponse.kt
@@ -1,3 +1,5 @@
 package fiofoundation.io.fiosdk.models.fionetworkprovider.response
 
-class GetFeeResponse(val fee: Long): FIOResponse()
+import java.math.BigInteger
+
+class GetFeeResponse(val fee: BigInteger): FIOResponse()
